Raise an exception when getting out of bounds includes
Fixes #5 
`json_include` specifying an index where nothing exists was the underlying problem. I have chosen to throw an error in that case, because nobody ought to be *trying* to get `nil` from that method. Since we are testing, a nice error message would probably be helpful.

I also added an error class so that expectations in this repo testing that an error is thrown when it should be are less likely to give a false positive.